Extract tensor object code to separate file
