+++ tensorrt-sys/trt-sys/TRTNetworkDefinition/TRTNetworkDefinition.cpp
@@ -4,8 +4,8 @@
 #include <NvInfer.h>
 
 #include "TRTNetworkDefinitionInternal.hpp"
-#include "../TRTDims/TRTDimsInternal.hpp"
 #include "../TRTLayer/TRTLayerInternal.hpp"
+#include "../TRTTensor/TRTTensorInternal.hpp"
 
 void destroy_network(Network_t *network) {
     if (network == nullptr)
@@ -19,10 +19,6 @@ nvinfer1::INetworkDefinition& Network::getNetworkDefinition() const {
 }
 
 Tensor_t *network_get_input(Network_t *network, int32_t idx) {
-    if (network == nullptr) {
-        return nullptr;
-    }
-
     return new Tensor(network->internal_network->getInput(idx));
 }
 
@@ -33,10 +29,3 @@ Layer_t* network_get_layer(Network_t *network, int index) {
     return new Layer(network->internal_network->getLayer(index));
 }
 
-void tensor_set_dimensions(Tensor_t *tensor, Dims_t *dimensions) {
-    if (tensor == nullptr) {
-        return;
-    }
-
-    tensor->internal_tensor->setDimensions(dims_get(dimensions));
-}
