+++ tensorrt-sys/trt-sys/TRTNetworkDefinition/TRTNetworkDefinition.h
@@ -5,28 +5,23 @@
 #ifndef LIBTRT_TRTNETWORKDEFINITION_H
 #define LIBTRT_TRTNETWORKDEFINITION_H
 
+#include <stdint.h>
+
+#include "../TRTLayer/TRTLayer.h"
+#include "../TRTTensor/TRTTensor.h"
+
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-#include <stdint.h>
-
-#include "../TRTDims/TRTDims.h"
-#include "../TRTLayer/TRTLayer.h"
-
 struct Network;
 typedef struct Network Network_t;
-
-struct Tensor;
-typedef struct Tensor Tensor_t;
 
 void destroy_network(Network_t *network);
 
 Layer_t* network_get_layer(Network_t *network, int index);
 
 Tensor_t *network_get_input(Network_t *network, int32_t idx);
-
-void tensor_set_dimensions(Tensor_t *tensor, Dims_t *dimensions); // only valid for input tensors
 
 #ifdef __cplusplus
 };
